# some changes to make performance better