--- a/index.html
+++ b/index.html
@@ -404,7 +404,6 @@
 
                                         <script src="static/js/vendor/modernizr-3.5.0.min.js"></script>
                                         <script src="https://code.jquery.com/jquery-3.2.1.min.js" integrity="sha256-hwg4gsxgFZhOsEEamdOYGBf13FyQuiTwlAQgxVSNgt4=" crossorigin="anonymous"></script>
-                                        <script>window.jQuery || document.write('<script src="static/js/vendor/jquery-3.2.1.min.js"><\/script>')</script>
                                         <script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.min.js"></script>
                                         <script src="static/js/plugins.js"></script>
                                         <script src="static/js/main.js"></script>