--- a/index.html
+++ b/index.html
@@ -407,7 +407,6 @@
                                         <script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.min.js"></script>
                                         <script src="static/js/plugins.js"></script>
                                         <script src="static/js/main.js"></script>
-                                         <script src="https://maps.googleapis.com/maps/api/js?callback=myMap"></script> 
 
                                         <!-- Google Analytics: change UA-XXXXX-Y to be your site's ID. -->
                                         <script>