--- a/index.html
+++ b/index.html
@@ -402,11 +402,11 @@
 
 
 
-                                        <script src="static/js/vendor/modernizr-3.5.0.min.js"></script>
-                                        <script src="https://code.jquery.com/jquery-3.2.1.min.js" integrity="sha256-hwg4gsxgFZhOsEEamdOYGBf13FyQuiTwlAQgxVSNgt4=" crossorigin="anonymous"></script>
-                                        <script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.min.js"></script>
-                                        <script src="static/js/plugins.js"></script>
-                                        <script src="static/js/main.js"></script>
+                                        <script src="static/js/vendor/modernizr-3.5.0.min.js" async></script>
+                                        <script src="https://code.jquery.com/jquery-3.2.1.min.js" integrity="sha256-hwg4gsxgFZhOsEEamdOYGBf13FyQuiTwlAQgxVSNgt4=" crossorigin="anonymous" async></script>
+                                        <script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.min.js" async></script>
+                                        <script src="static/js/plugins.js" async></script>
+                                        <script src="static/js/main.js" async></script>
 
                                         <!-- Google Analytics: change UA-XXXXX-Y to be your site's ID. -->
                                         <script>
@@ -414,31 +414,7 @@
                                             ga('create','UA-72748005-1','auto');ga('send','pageview')
                                         </script>
                                         <script src="https://www.google-analytics.com/analytics.js" async defer></script>
-                                        <script type="text/javascript" src="https://cdnjs.cloudflare.com/ajax/libs/wow/1.1.2/wow.min.js"></script>
-                                        <script type="text/javascript">
-                                            $(window).scroll(function(){
-                                                if ($(window).scrollTop() >= 10) {
-                                                    $('nav').addClass('shadow-header');
-                                                }
-                                                else {
-                                                    $('nav').removeClass('shadow-header');
-                                                }
-                                            });
-                                        </script>
-                                        <script type="text/javascript" src="static/js/faq.js"></script>
-                                        <script>
-                                          new WOW().init();
-                                      </script>
-                                      <script type="text/javascript">
-                                        $('.faq li .question').click(function () {
-                                          $(this).find('.plus-minus-toggle').toggleClass('collapsed');
-                                          $(this).parent().toggleClass('active');
-                                      });
-                                  </script>
-                                  <script type="text/javascript">
-                                      $(document).ready(function(){
-                                        $(".navbar-default").css("background-color", "#fff");
-                                      });
-                                  </script>
+                                        <script type="text/javascript" src="https://cdnjs.cloudflare.com/ajax/libs/wow/1.1.2/wow.min.js" async></script>
+                                        <script type="text/javascript" src="static/js/faq.js" async></script>
                               </body>
                               </html>