--- a/index.html
+++ b/index.html
@@ -15,10 +15,10 @@
     <!-- TO DO : Remove unused fonts imports -->
     <link href="https://fonts.googleapis.com/css?family=Lato:400,700|Montserrat:400,700|Open+Sans:400,500,600,700|Oswald:400,700|Roboto:400,700" rel="stylesheet">
     <link href="https://fonts.googleapis.com/css?family=Vollkorn" rel="stylesheet">
-    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css">
-    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/font-awesome/4.7.0/css/font-awesome.min.css">
+    <link rel="stylesheet" href="static/css/bootstrap.min.css">
+    <link rel="stylesheet" href="static/css/font-awesome.min.css">
     <link rel="stylesheet" href="static/css/main.css">
-    <link rel="stylesheet" type="text/css" href="https://cdnjs.cloudflare.com/ajax/libs/animate.css/3.5.2/animate.min.css">
+    <link rel="stylesheet" type="text/css" href="static/css/animate.min.css">
     <link href="static/css/hover-min.css" rel="stylesheet" media="all">
     <link rel="stylesheet" type="text/css" href="static/css/cards.css">
 </head>
@@ -403,8 +403,8 @@
 
 
                                         <script src="static/js/vendor/modernizr-3.5.0.min.js" async></script>
-                                        <script src="https://code.jquery.com/jquery-3.2.1.min.js" integrity="sha256-hwg4gsxgFZhOsEEamdOYGBf13FyQuiTwlAQgxVSNgt4=" crossorigin="anonymous" async></script>
-                                        <script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.min.js" async></script>
+                                        <script src="static/js/vendor/jquery-3.2.1.min.js" async></script>
+                                        <script src="static/js/bootstrap.min.js" async></script>
                                         <script src="static/js/plugins.js" async></script>
                                         <script src="static/js/main.js" async></script>
 
@@ -413,8 +413,8 @@
                                             window.ga=function(){ga.q.push(arguments)};ga.q=[];ga.l=+new Date;
                                             ga('create','UA-72748005-1','auto');ga('send','pageview')
                                         </script>
-                                        <script src="https://www.google-analytics.com/analytics.js" async defer></script>
-                                        <script type="text/javascript" src="https://cdnjs.cloudflare.com/ajax/libs/wow/1.1.2/wow.min.js" async></script>
+                                        <script src="static/js/analytics.js" async defer></script>
+                                        <script type="text/javascript" src="static/js/wow.min.js" async></script>
                                         <script type="text/javascript" src="static/js/faq.js" async></script>
                               </body>
                               </html>