… static/css/animate.min.css
… static/css/bootstrap.min.css
… static/css/font-awesome.min.css
… static/js/analytics.js