removed inline javascripts, shifted them in main js, added async attribute. Noticed site loading faster from 1.02s to 440 ms in local pc
on github pages: (monitored on slow 3G connection)
on the screenshot posted on fb it was 40s for total and 25s for DOM Content
now it is 25s for total and 5s for DOM content.

It's 9s total on my normal wifi connection

![screenshot from 2017-12-07 14-35-41](https://user-images.githubusercontent.com/21174572/33707097-40d9b92e-db5c-11e7-8f21-d9728e1ebb29.png)